Permission backend

+++ app/Http/Controllers/Api/PostController.php
@@ -52,6 +52,7 @@ class PostController extends Controller
 
     public function store(StorePostRequest $request)
     {
+        $this->authorize('posts.create');
         if ($request->hasFile('thumbnail')) {
             $filename = $request->file('thumbnail')->getClientOriginalName();
             info($filename);
@@ -64,16 +65,20 @@ class PostController extends Controller
 
     public function show(Post $post)
     {
+        $this->authorize('posts.update');
         return new PostResource($post);
     }
 
     public function update(Post $post, UpdatePostRequest $request)
     {
+        $this->authorize('posts.update');
         $post->update($request->validated());
         return new PostResource($post);
     }
 
     public function destroy(Post $post) {
+
+        $this->authorize('posts.delete');
         $post->delete();
         return response()->noContent();
     }
